Parametro de mês e valor inicial Dados do swagger

+++ API Calculadora/BancoACME.Interface/ICalculadora.cs
@@ -5,6 +5,6 @@ namespace BancoACME.Interface
 {
     public interface ICalculadora
     {
-        Task<double> calculaJuros(double valorInicial, double meses);
+        Task<double> calculaJuros(double valorInicial, int meses);
     }
 }
